don't eat "\" in the incoming json
Before this commit, the post-receive.cgi script would erroneously
interpret the slash character \ in the incoming json POST message.

This caused problems when commit messages contained ", because the
github webhook would escape the " and the cgi script ould eat the escape
character, and the resulting string would not be valid json.