tweak heading sizes
Way above, there are settings for "h1", "h2" and so forth in percentage
font sizes.  Down here, there are more specific rules for "h2" inside
".post-content" and so forth.  Originally, there was no rule for
.post-content h1 and in practice, h1 was smaller than .post-content h2.

Probably it would be better to get rid of some of the redundant rules,
but for now just paper over it.